% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbCfPred}
\alias{pbCfPred}
\title{Predict for posterior samples from the interim analysis}
\usage{
pbCfPred(vq, vp, n2, theta0, alpha, repeach)
}
\description{
Predict for posterior samples from the interim analysis
}
