% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbCfPredSingle}
\alias{pbCfPredSingle}
\title{Predict for a single set of value for rep times}
\usage{
pbCfPredSingle(vq, vp, n2, theta0, alpha, rep, rst)
}
\arguments{
\item{rep}{number of replications}
}
\description{
Predict for a single set of value for rep times
}
