% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbCfYtest}
\alias{pbCfYtest}
\title{Final Hypothesis testing}
\usage{
pbCfYtest(y, theta0)
}
\arguments{
\item{y}{observed binary outcome data}

\item{theta0}{reference theta for H0}
}
\description{
Final Hypothesis testing
}
