% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sampling.R
\name{pbCumuPQ}
\alias{pbCumuPQ}
\title{Posterior cumulative response rate}
\usage{
pbCumuPQ(post.q, post.p)
}
\arguments{
\item{post.q}{biomarker interval probabilities}

\item{post.p}{response rates}
}
\value{
biomarker probabilities and response rates conditioning on biomarker
    values bigger than cut points
}
\description{
Posterior cumulative response rate
}
