% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_trial.R
\name{pbSimuSingleTrial}
\alias{pbSimuSingleTrial}
\title{Simulate a single trial}
\usage{
pbSimuSingleTrial(par.biom, par.resp, n2, theta0, prior.p = NULL,
  prior.q = c(a = 0.5, b = 0.5), cut.x = NULL, cut.quants = c(0.25,
  0.5, 0.75), cut.known = TRUE, cand.cuts = 1:(length(cut.quants) + 1),
  resp.mdl = "simplebin", uti.f, alpha = 0.05, uti.cut = 0.1,
  B1 = 1, C1 = 1, C2 = 0, C3 = 0, iter = 4000, nlarge = 50000,
  repeach = 1)
}
\arguments{
\item{par.biom}{list of parameters for simulating biomarkers}

\item{par.resp}{list of parameters for simulating responses}

\item{iter}{number of iterations for interim analysis}

\item{nlarge}{large n for evaluating the truth}
}
\description{
Simulate a single trial
}
