% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sampling.R
\name{pbSmpBiom}
\alias{pbSmpBiom}
\title{Posterior distribution of biomarker as intervals}
\usage{
pbSmpBiom(x.count, prior.p = NULL, iter = 4000)
}
\description{
Posterior distribution of biomarker as intervals
}
