% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_utility.R
\name{pbUtility}
\alias{pbUtility}
\title{Compute utility}
\usage{
pbUtility(uti, ...)
}
\description{
theta0      <- ["theta0"];
resp.all    <- ["resp.all"];
resp2       <- ["resp2"];
r           <- ["Stage1prop"];
est.t       <- ["est.t"];
etheta      <- ["etheta"];
rej         <- ["rej"];
n.ScreenOut <- ["n.ScreenOut"];
n1          <- ["n1"]; 	n2   <- rst.summary["n2"];
p.extra     <- n.ScreenOut/(n1+n2)
p.resp.all  <- resp.all/(n1+n2)
p.resp2     <- resp2/(n1+n2)
}
