% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sampling.R
\name{prvSmpSimplebin}
\alias{prvSmpSimplebin}
\title{Private Function: Sampling from simple binomial of response rates}
\usage{
prvSmpSimplebin(x.cut, y, cand.cuts, iter = 4000, prior.q = c(a = 0.5,
  b = 0.5), ...)
}
\description{
Private Function: Sampling from simple binomial of response rates
}
