% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Random draw from Dirichlet}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of samples}

\item{alpha}{dirichlet parameters}
}
\description{
Random draw from Dirichlet
}
