% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tkListAddValue}
\alias{tkListAddValue}
\title{Set default values in a list}
\usage{
tkListAddValue(tar.lst, value.lst, overwrite = FALSE)
}
\arguments{
\item{tar.lst}{target list}

\item{value.lst}{list with values}

\item{overwrite}{Whether overwrite tar.lst with value.lst}
}
\description{
Set default values in a list
}
