% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbiom-package.R
\docType{package}
\name{pbiom-package}
\alias{pbiom-package}
\alias{pbiom}
\title{The 'pbiom' package.}
\description{
Phase II Biomarker Study
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
