% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_trial.R
\name{pbSimuSingleTrial}
\alias{pbSimuSingleTrial}
\title{Simulate a single trial}
\usage{
pbSimuSingleTrial(par.biom, par.resp, n2, prior.p = NULL, prior.q = c(a
  = 0.5, b = 0.5), theta0, uti.f, alpha = 0.05, cuts = NULL,
  probs = c(0.25, 0.5, 0.75), cand.cuts = 1:(length(probs) + 1),
  resp.mdl = "simplebin", iter = 4000, nlarge = 10000, repeach = 1)
}
\arguments{
\item{par.biom}{list of parameters for simulating biomarkers}

\item{par.resp}{list of parameters for simulating responses}

\item{iter}{number of iterations for interim analysis}

\item{nlarge}{large n for evaluating the truth}
}
\description{
Simulate a single trial
}
