% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_sampling.R
\name{pbSmpResp}
\alias{pbSmpResp}
\title{Posterior response rates for each cut interval}
\usage{
pbSmpResp(x.cut, y, cand.cuts = NULL, type = c("simplebin"),
  iter = 4000, ...)
}
\arguments{
\item{x.cut}{observed biomarker levels}

\item{y}{response status}

\item{cand.cuts}{candidate cut points}
}
\description{
Posterior response rates for each cut interval
}
